Add Request Manager
