+++ utecio/latch5nfc.py
@@ -2,22 +2,21 @@ import asyncio
 
 from __init__ import logger
 from lock import UtecBleLock
-from enums import BleResponseCode, BLECommandCode, ServiceUUID
-from device import BleDeviceKeyECC, BleRequest, BleResponse
+from enums import BLECommandCode, ULDeviceModel
+from device import BleRequest
 from constants import Latch5_NFC, BLE_RETRY_DELAY_DEF, BLE_RETRY_MAX_DEF
 
 class Latch5NFC(UtecBleLock):
-    def __init__(self,  uid: str, password: str, mac_uuid: str, device_name: str = Latch5_NFC, wurx_uuid: str = None, max_retries: float = BLE_RETRY_MAX_DEF, retry_delay: float = BLE_RETRY_DELAY_DEF, bleakdevice_callback: callable = None):
+    def __init__(self,  uid: str, password: str, mac_uuid: str, device_name: str = Latch5_NFC, wurx_uuid: str = None, max_retries: float = BLE_RETRY_MAX_DEF, retry_delay: float = BLE_RETRY_DELAY_DEF):
         super().__init__(uid=uid, 
                          password=password, 
                          mac_uuid=mac_uuid,
                          wurx_uuid=wurx_uuid,
                          device_name=device_name, 
                          max_retries=max_retries, 
-                         retry_delay=retry_delay, 
-                         bleakdevice_callback=bleakdevice_callback)
+                         retry_delay=retry_delay)
         
-        self.key = BleDeviceKeyECC()
+        self.model = ULDeviceModel.L5NFC.value
         self.capabilities.bluetooth = True
         self.capabilities.autolock = True
         self.capabilities.update_wifi = True
@@ -41,13 +40,12 @@ class Latch5NFC(UtecBleLock):
 
     async def update(self):
         try:
-            await self.start_notify(ServiceUUID.DATA.value, self._receive_write_response)
-            await self.send_encrypted(BleRequest(BLECommandCode.LOCK_STATUS))
-            await asyncio.sleep(2)
-            await self.stop_notify(ServiceUUID.DATA.value)
-            logger.debug(f"({self.client.address}) Update request completed.")
+            self.queue_request(BleRequest(command=BLECommandCode.LOCK_STATUS))
+            await self.process_queue()
+            
+            logger.debug(f"({self.mac_address}) Update request completed.")
         except Exception as e:
-            logger.error(f"({self.client.address}) Error during update request: {e}")
+            logger.error(f"({self.mac_address}) Error during update request: {e}")
             
         return await super().update()
