+++ utecio/ul1bt.py
@@ -2,8 +2,8 @@ import asyncio
 
 from __init__ import logger
 from lock import UtecBleLock
-from enums import BLECommandCode, ServiceUUID, ULDeviceModel
-from device import BleDeviceKeyMD5, BleRequest
+from enums import BLECommandCode, ULDeviceModel
+from device import BleRequest
 from constants import UL1_BT, BLE_RETRY_DELAY_DEF, BLE_RETRY_MAX_DEF
 
 class UL1BT(UtecBleLock):
@@ -17,8 +17,7 @@ class UL1BT(UtecBleLock):
                          retry_delay=retry_delay, 
                          bleakdevice_callback=bleakdevice_callback)
         
-        self.key = BleDeviceKeyMD5()
-        self.model = ULDeviceModel.UL1BT
+        self.model = ULDeviceModel.UL1BT.value
         self.capabilities.bluetooth = True
         self.capabilities.rfid = True
         self.capabilities.rfid_twice = True
@@ -39,15 +38,14 @@ class UL1BT(UtecBleLock):
 
     async def update(self):
         try:
-            await self.start_notify(ServiceUUID.DATA.value, self._receive_write_response)
-            await self.send_encrypted(BleRequest(BLECommandCode.GET_LOCK_STATUS))
-            await self.send_encrypted(BleRequest(BLECommandCode.GET_BATTERY))
-            await self.send_encrypted(BleRequest(BLECommandCode.GET_SN, None, None, bytearray([16])))
-            await self.send_encrypted(BleRequest(BLECommandCode.GET_MUTE))
-            await asyncio.sleep(2)
-            await self.stop_notify(ServiceUUID.DATA.value)
-            logger.debug(f"({self.client.address}) Update request completed.")
+            self.queue_request(BleRequest(command=BLECommandCode.GET_LOCK_STATUS))
+            self.queue_request(BleRequest(command=BLECommandCode.GET_BATTERY))
+            self.queue_request(BleRequest(command=BLECommandCode.GET_SN, data=bytearray([16])))
+            self.queue_request(BleRequest(command=BLECommandCode.GET_MUTE))
+            await self.process_queue()    
+
+            logger.debug(f"({self.mac_address}) Update request completed.")
         except Exception as e:
-            logger.error(f"({self.client.address}) Error during update request: {e}")
+            logger.error(f"({self.mac_address}) Error during update request: {e}")
             
         return await super().update()
